Remove gender from Toons
Requires further testing, since estates cannot be loaded currently.
This replaces the 'gender' DNA field with eyelashes, and removes
gender-specific clothing and accessories restrictions. This aims
to be backwards-compatible with the original TTO DNAs, but the new
DNAs that are generated are not compatible with the original code.
Keep this in mind when running the game with this branch.