Allow choosing accessories from the clothes GUI
This allows the avatar to browse and buy accessories from a tailor.

The Make-A-Toon screen also permits a new avatar to wear hats and
shoes alongside clothing. Some backend work was done to make that
possible. Additionally, the avatar selection screen can now render
the accessories the player was wearing on their avatar.